moving replay buffer and target net mixins into their respective modules
