+++ pandemonium/demons/offline_td.py
@@ -1,11 +1,8 @@
-from copy import deepcopy
-
 import torch
 import torch.nn.functional as F
+from pandemonium.demons import Demon, Loss, ControlDemon, PredictionDemon
+from pandemonium.experience import (Trajectory, Transitions)
 
-from pandemonium.demons import (Demon, Loss, ControlDemon, PredictionDemon,
-                                ParametricDemon)
-from pandemonium.experience import ER, PER, Trajectory, Transitions
 
 
 class OfflineTD(Demon):
@@ -137,117 +134,3 @@ class TDn(TTD):
             eligibility=lambda trajectory: torch.ones_like(trajectory.r),
             **kwargs
         )
-
-
-class ReplayBufferMixin:
-    r""" Mixin that adds a replay buffer to an agent.
-
-    Was originally designed as a means to make RL more data efficient [1].
-    Later on adapted in DQN architecture to make the data distribution more
-    stationary [2].
-
-    References
-    ----------
-    [1] "Self-Improving Reactive Agents Based On RL, Planning and Teaching"
-        by Lin. http://www.incompleteideas.net/lin-92.pdf
-    [2] "Playing Atari with Deep Reinforcement Learning"
-        https://arxiv.org/pdf/1312.5602.pdf
-    """
-    delta: callable
-
-    def __init__(self,
-                 replay_buffer: ER,
-                 priority_measure: str = 'td_error'):
-        """
-
-        Parameters
-        ----------
-        replay_buffer: ReplayBuffer
-            An instance of a replay buffer
-        priority_measure: str
-            Priorities to experiences are assigned based on this metric in
-            prioritized ER case. Defaults to using `td_error`, but could also
-            be `ce_loss` in case of distributional value learning.
-        """
-        self.replay_buffer = replay_buffer
-        self.priority_measure = priority_measure
-
-    def store(self, transitions: Transitions):
-        """ Adds transitions to the replay buffer.
-
-        Pre-computes priorities in case Prioritized Experience Replay is used.
-        """
-
-        if isinstance(self.replay_buffer, PER):
-            trajectory = Trajectory.from_transitions(transitions)
-            _, info = self.delta(trajectory)
-            priorities = info[self.priority_measure]
-            priorities = priorities.abs() + self.replay_buffer.ε
-            self.replay_buffer.add_batch(transitions, priorities.tolist())
-        else:
-            self.replay_buffer.add_batch(transitions)
-
-    def _update_priorities(self, trajectory: Trajectory, info: dict):
-        priorities = info[self.priority_measure]
-        priorities = priorities.abs() + self.replay_buffer.ε
-        indexes = trajectory.buffer_index.tolist()
-        self.replay_buffer.update_priorities(indexes, priorities.tolist())
-
-
-class TargetNetMixin:
-    r""" Mixin that adds a target network to the agent.
-
-    Duplicates the estimator networks that are used to estimate targets.
-    These clones are updated at a `target_update_freq` frequency which allows
-    to stabilize the training process and make targets less non-stationary.
-
-    References
-    ----------
-    [1] "Playing Atari with Deep Reinforcement Learning"
-        https://arxiv.org/pdf/1312.5602.pdf
-    """
-    aqf: callable
-    avf: callable
-    φ: callable
-
-    def __init__(self, target_update_freq: int = 0):
-
-        self.update_counter = 0  # keeps track of the number of updates so far
-
-        # Create target networks to stabilize training
-        self.target_networks = dict()
-        self.target_update_freq = target_update_freq
-        if self.target_update_freq:
-            self.target_feature = deepcopy(self.φ)
-            self.target_feature.load_state_dict(self.φ.state_dict())
-            self.target_feature.eval()
-            self.target_networks[self.target_feature] = self.φ
-
-            if isinstance(self.avf, nn.Module):
-                # `avf` is a network in duelling DQN and is implicit in `aqf`
-                # in other cases
-                self.target_avf = deepcopy(self.avf)
-                self.target_avf.load_state_dict(self.avf.state_dict())
-                self.target_avf.eval()
-                self.target_networks[self.target_avf] = self.avf
-
-            if isinstance(self.aqf, nn.Module):
-                # `aqf` is a network in all DQN family except distributional
-                # case, where it is implicit in the `azf`
-                self.target_aqf = deepcopy(self.aqf)
-                self.target_aqf.load_state_dict(self.aqf.state_dict())
-                self.target_aqf.eval()
-                self.target_networks[self.target_aqf] = self.aqf
-
-            if hasattr(self, 'azf') and isinstance(self.azf, nn.Module):
-                # only for distributional agents
-                self.target_azf = deepcopy(self.azf)
-                self.target_azf.load_state_dict(self.azf.state_dict())
-                self.target_azf.eval()
-                self.target_networks[self.target_azf] = self.azf
-
-    def sync_target(self):
-        if self.target_update_freq:
-            if self.update_counter % self.target_update_freq == 0:
-                for target_net, net in self.target_networks.items():
-                    target_net.load_state_dict(net.state_dict())
